Pulled C2Connection out into its own class file. Cleaned up the class and made it a bit more .NET
