+++ SnareNetStd/SnareNetStd.csproj
@@ -47,9 +47,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="C2Connection.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="screenshot.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
