--- SnareNetStd/screenshot.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Drawing.Imaging;
-using System.Drawing;
-using System.Windows.Forms;
-using System.IO;
-
-namespace SnareNetStd
-{
-    public class Screenshot
-    {
-        public Image ScreenShot()
-        {
-            throw new NotImplementedException();
-        }
-
-
-
-    }
-}
